+++ docs/changelog.rst
@@ -11,6 +11,7 @@ version 0.7.0.
 Unreleased
 ----------
 
+- ADD: `~cernml.lsa_utils.trim_scalar_settings()`.
 - FIX: Several broken links in the documentation.
 - OTHER: Switched project manifest from ``setup.cfg`` to ``pyproject.toml``.